% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: reinforced regression}
\description{
lists of data sets capable & adjustable clusters.
  regression and confidence interval.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
