% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{object of formula}

\item{data}{a data set}

\item{n}{overall sample size}
}
\description{
compute the regression estimates for a blb dataset
}
