% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{linear regression based on one subsample}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a formula class object}

\item{data}{data frame for analysis}

\item{n}{sample size}

\item{B}{bootstrap iteration times}
}
\description{
linear regression based on one subsample
}
