% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{a function to split dataframe}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{name of dataset}

\item{m}{split to m parts
split data into m parts of approximated equal sizes}
}
\description{
a function to split dataframe
}
