% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B, cpu)
}
\arguments{
\item{formula}{object of formula ex. x~y}

\item{data}{name of dataset}

\item{n}{sample size}

\item{B}{bootstrap size}

\item{cpu}{number of cpu}
}
\description{
compute the estimates
}
